Allow for other headers to be redacted
Since Req already redacts "authorization" headers automatically, it
seems like not much more of a stretch to allow for other headers to be
redacted.

This commit allows users to use the application environment to configure
additional headers that should be redacted in the log output. This
option is exposed at the global level to allow configuration of the
global logging implementation.